Working towards 0.9.2.0
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@311 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ mmex/src/dbwrapper.cpp
@@ -499,15 +499,21 @@ void removeCruft(wxSQLite3Database* db)
     mmBEGINSQL_LITE_EXCEPTION;
 
     {
-        wxSQLite3StatementBuffer bufSQL3;
-        bufSQL3.Format("DELETE FROM SPLITTRANSACTIONS_V1 WHERE SPLITTRANSACTIONS_V1.TRANSID NOT IN (SELECT TRANSID FROM CHECKINGACCOUNT_V1);");
-        db->ExecuteUpdate(bufSQL3);
+        if (db->TableExists(wxT("SPLITTRANSACTIONS_V1")))
+        {
+            wxSQLite3StatementBuffer bufSQL3;
+            bufSQL3.Format("DELETE FROM SPLITTRANSACTIONS_V1 WHERE SPLITTRANSACTIONS_V1.TRANSID NOT IN (SELECT TRANSID FROM CHECKINGACCOUNT_V1);");
+            db->ExecuteUpdate(bufSQL3);
+        }
     }
 
     {
-        wxSQLite3StatementBuffer bufSQL3;
-        bufSQL3.Format("DELETE FROM BUDGETSPLITTRANSACTIONS_V1 WHERE BUDGETSPLITTRANSACTIONS_V1.TRANSID NOT IN (SELECT BDID FROM BILLSDEPOSITS_V1);");
-        db->ExecuteUpdate(bufSQL3);
+        if (db->TableExists(wxT("BUDGETSPLITTRANSACTIONS_V1")))
+        {
+            wxSQLite3StatementBuffer bufSQL3;
+            bufSQL3.Format("DELETE FROM BUDGETSPLITTRANSACTIONS_V1 WHERE BUDGETSPLITTRANSACTIONS_V1.TRANSID NOT IN (SELECT BDID FROM BILLSDEPOSITS_V1);");
+            db->ExecuteUpdate(bufSQL3);
+        }
     }
 
 #if 0
